+++ betfairlightweight/streaming/stream.py
@@ -83,7 +83,7 @@ class BaseStream:
 
     def snap(self, market_ids: list = None) -> list:
         return [
-            cache.create_resource(self.unique_id, self._lightweight)
+            cache.create_resource(self.unique_id, self._lightweight, snap=True)
             for cache in list(self._caches.values())
             if market_ids is None or cache.market_id in market_ids
         ]